Improve missing arg handling and titles in plots
Titles: Now show three arguments per line, so that it's not a single really
long line when lots of arguments are specified in title_specs.

Missing argument handling: This removes the momentum_client hack. For this
to work properly, I needed to move directories that don't have the
'momentum_client' option to a new directory, and exclude that directory from
the plots that used momentum_client=0.9. This is because the new mechanism
doesn't treat a missing momentum_client as if it were 0.0---this now needs to
be specified as [0.0, '__missing__'] if that's what's intended. It's probably
better to be explicit about this.

I checked all of the plots in the updated notebooks visually against the old
versions to make sure they looked the same with the same sample sizes.